+++ test/unit/agent.test.js
@@ -1,13 +1,29 @@
 'use strict'
 
+const Fs = require('fs')
+const Path = require('path')
 const Http = require('http')
 const Https = require('https')
 const Tls = require('tls')
+const Sinon = require('sinon')
 const { assert } = require('chai')
 
 const Agent = require('../../lib/agent')
 
 describe('agents', () => {
+  let sandbox
+
+  beforeEach(() => {
+    sandbox = Sinon.createSandbox()
+    sandbox.origGlobalAgentOptions = { ...Https.globalAgent.options }
+  })
+
+  afterEach(() => {
+    Https.globalAgent.options = sandbox.origGlobalAgentOptions
+    sandbox.restore()
+    sandbox = null
+  })
+
   it('should return an http agent', () => {
     const instance = Agent.create('http')
 
@@ -46,6 +62,8 @@ describe('agents', () => {
   })
 
   it('should return an https agent using the provided secureContextOptions', () => {
+    sandbox.spy(Tls, 'createSecureContext')
+
     const instance = Agent.create('https', {
       agentOptions: {
         secureContextOptions: {
@@ -56,5 +74,34 @@ describe('agents', () => {
 
     assert.ok(instance instanceof Https.Agent, 'is https agent.')
     assert.ok(instance.options.secureContext, 'is pre-cached secureContext.')
+    Sinon.assert.calledWith(Tls.createSecureContext, { ca: ['cert1'], cert: undefined, key: undefined })
+  })
+
+  it('should include globalAgent options: ca, cert, key', () => {
+    sandbox.spy(Tls, 'createSecureContext')
+
+    const ca = Fs.readFileSync(Path.join(__dirname, '..', 'fixtures', 'ca.crt'))
+    const cert = Fs.readFileSync(Path.join(__dirname, '..', 'fixtures', 'cert.pem'))
+    const key = Fs.readFileSync(Path.join(__dirname, '..', 'fixtures', 'key.pem'))
+
+    Https.globalAgent.options = { ca: [ca], cert, key }
+
+    const instance = Agent.create('https', {
+      agentOptions: {
+        secureContextOptions: {
+          foo: 'bar'
+        }
+      }
+    })
+
+    assert.ok(instance instanceof Https.Agent, 'is https agent.')
+    assert.ok(instance.options.secureContext, 'is pre-cached secureContext.')
+
+    const options = Tls.createSecureContext.args[0][0]
+
+    assert.ok(options.cert === cert, 'has global agent cert')
+    assert.ok(options.key === key, 'has global agent key')
+    assert.ok(options.ca.length > 10, 'has root certificates')
+    assert.ok(options.ca.includes(ca), 'has global agent ca')
   })
 })